Add implementation of collect_runs() and collect_1q_runs()
This commit adds a rust implementation of the collect runs and collect
1q runs functions to the DAGCircuit class. It adds a native rust api
that returns an iterator of node indices and also a python interface
that matches the expectations from the previous python implementation.

The one difference is that the return type for collect_1q_runs() will
change from an `rx.NodeIndices` to a Python list. This isn't a huge deal
in practice as the NodeIndices is just a custom sequence type and the
python list should work identically except for explicit type checks.
Although, this actually matches the Python type hint of the previous
implementation. There wasn't really an alternative for this because
rustworkx doesn't expose an interface to create NodeIndices, they're
read only return types.