Add implementation of collect_2q_runs()
This commit adds a rust implementation of the collect 2q runs method
to the DAGCircuit class. It adds a native rust api that returns an
vec of vec of node indices and also a python interface that matches the
expectations from the previous python implementation. The rustworkx-core
function returns a collected vec so instead of throwing that away for an
iterator this just returns the owned vec so the callers can decide how
to use it.

The one difference is that the return type for collect_1q_runs() will
change from an `rx.NodeIndices` to a Python list. This isn't a huge deal
in practice as the NodeIndices is just a custom sequence type and the
python list should work identically except for explicit type checks.