Implement dag.draw() method
This commit implements the dag.draw() method for the new rust dagcircuit
struct. To accomplish this first the dot_utils.rs module from rustworkx
is forked and included in qiskit circuit crate. This module is used to
generate a dot file using python callbacks, it is modified to work with
the dagcircuit types directly. Eventually this can be ported to
rustworkx-core and made more generic, but honestly there isn't too much
of a need because that will require a lot of boiler plate on the caller
side here. With a method to generate a dot file from rust for a
dagcircuit implement this is then exposed to Python as a new private
method `._to_dot()` which will return a dot file string. The Python
space dag drawer is then modified to instead of relying on rustworkx's
graphviz_draw() method to instead call `dag._to_dot()` and manually pass
that to graphviz to generate a visualization. This is essentially what
rustworkx's graphviz_draw() is already doing under the covers. Then the
final piece to implement the rust implementation of the draw() method is
to use pyo3 to call the python space dag_drawer() function which is how
the previous python implementation's method worked.